backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyyaml==6.0.2
pydantic==2.10.6
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
langgraph==0.2.72
langchain==0.3.14
langchain-groq==0.2.4
langchain-core==0.3.28
